Split up TextLink view and its enums
